ble_app/requirements.txt
bleak==0.22.2
dbus-fast==2.22.1
pycryptodome==3.20.0
typing_extensions==4.12.2
